Extract shared port-to-PID map, add platform-specific remote tag resolution
Move getListeningPortToPIDMap to net_portmap.go (linux || windows) to
eliminate duplication. Add getRemoteProcessTags as a platform-specific
function: Windows uses process cache from system-probe, Linux uses the
tagger. Skip remote tag resolution for containerized connections.
