+++ pkg/process/checks/net_linux.go
@@ -13,27 +13,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/process/net"
 	"github.com/DataDog/datadog-agent/pkg/util/cloudproviders/network"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
-	"github.com/DataDog/datadog-agent/pkg/util/port/portlist"
 )
-
-// getListeningPortToPIDMap returns a map of listening port -> PID using the portlist Poller
-func getListeningPortToPIDMap() map[int32]int32 {
-	poller := &portlist.Poller{IncludeLocalhost: true}
-	defer poller.Close()
-
-	ports, _, err := poller.Poll()
-	if err != nil {
-		log.Debugf("failed to poll listening ports: %v", err)
-		return nil
-	}
-	result := make(map[int32]int32, len(ports))
-	for _, p := range ports {
-		if p.Pid > 0 {
-			result[int32(p.Port)] = int32(p.Pid)
-		}
-	}
-	return result
-}
 
 // fetchIISTagsCache is not applicable on Linux; returns nil.
 func fetchIISTagsCache(_ *http.Client) map[string][]string {
@@ -43,6 +23,19 @@ func fetchIISTagsCache(_ *http.Client) map[string][]string {
 // fetchProcessCacheTags is not applicable on Linux; returns nil.
 func fetchProcessCacheTags(_ *http.Client) map[uint32][]string {
 	return nil
+}
+
+// getRemoteProcessTags returns process tags for a remote PID using the tagger.
+func getRemoteProcessTags(pid int32, _ map[uint32][]string, processTagProvider func(int32) ([]string, error)) []string {
+	if processTagProvider == nil {
+		return nil
+	}
+	tags, err := processTagProvider(pid)
+	if err != nil {
+		log.Debugf("error getting process tags for remote pid %d: %v", pid, err)
+		return nil
+	}
+	return tags
 }
 
 // getNetworkID fetches network_id from the current netNS or from the system probe if necessary, where the root netNS is used
