+++ pkg/process/checks/net_windows.go
@@ -14,31 +14,11 @@ import (
 	sysconfig "github.com/DataDog/datadog-agent/pkg/system-probe/config"
 	"github.com/DataDog/datadog-agent/pkg/util/cloudproviders/network"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
-	"github.com/DataDog/datadog-agent/pkg/util/port/portlist"
 )
 
 // getNetworkID fetches network_id
 func getNetworkID(_ *http.Client) (string, error) {
 	return network.GetNetworkID(context.Background())
-}
-
-// getListeningPortToPIDMap returns a map of listening port -> PID using the portlist Poller
-func getListeningPortToPIDMap() map[int32]int32 {
-	poller := &portlist.Poller{IncludeLocalhost: true}
-	defer poller.Close()
-
-	ports, _, err := poller.Poll()
-	if err != nil {
-		log.Debugf("failed to poll listening ports: %v", err)
-		return nil
-	}
-	result := make(map[int32]int32, len(ports))
-	for _, p := range ports {
-		if p.Pid > 0 {
-			result[int32(p.Port)] = int32(p.Pid)
-		}
-	}
-	return result
 }
 
 // fetchIISTagsCache retrieves the IIS tags cache from system-probe's /iis_tags endpoint.
@@ -93,4 +73,12 @@ func fetchProcessCacheTags(client *http.Client) map[uint32][]string {
 		return nil
 	}
 	return result
+}
+
+// getRemoteProcessTags returns process tags for a remote PID using the system-probe process cache.
+func getRemoteProcessTags(pid int32, procCacheTags map[uint32][]string, _ func(int32) ([]string, error)) []string {
+	if procCacheTags == nil {
+		return nil
+	}
+	return procCacheTags[uint32(pid)]
 }
